+++ tests/test_internal_histogram.py
@@ -4,7 +4,7 @@ import boost_histogram as bh
 import numpy as np
 from numpy.testing import assert_array_equal, assert_allclose
 
-methods = (bh.storage.Double, bh.storage.Int, bh.storage.Unlimited)
+methods = (bh.storage.Double, bh.storage.Int64, bh.storage.Unlimited)
 
 
 @pytest.mark.parametrize("dtype", [np.double, np.int_, np.float32])
@@ -69,7 +69,7 @@ def test_2D_fill_int(storage):
 
 def test_edges_histogram():
     edges = (1, 12, 22, 79)
-    hist = bh.Histogram(bh.axis.Variable(edges), storage=bh.storage.Int())
+    hist = bh.Histogram(bh.axis.Variable(edges), storage=bh.storage.Int64())
 
     vals = (13, 15, 24, 29)
     hist.fill(vals)
@@ -81,7 +81,7 @@ def test_edges_histogram():
 
 
 def test_int_histogram():
-    hist = bh.Histogram(bh.axis.Integer(3, 7), storage=bh.storage.Int())
+    hist = bh.Histogram(bh.axis.Integer(3, 7), storage=bh.storage.Int64())
 
     vals = (1, 2, 3, 4, 5, 6, 7, 8, 9)
     hist.fill(vals)
@@ -93,7 +93,9 @@ def test_int_histogram():
 
 
 def test_str_categories_histogram():
-    hist = bh.Histogram(bh.axis.StrCategory(["a", "b", "c"]), storage=bh.storage.Int())
+    hist = bh.Histogram(
+        bh.axis.StrCategory(["a", "b", "c"]), storage=bh.storage.Int64()
+    )
 
     vals = ["a", "b", "b", "c"]
     # Can't fill yet
@@ -101,7 +103,7 @@ def test_str_categories_histogram():
 
 def test_growing_histogram():
     hist = bh.Histogram(
-        bh.axis.Regular(10, 0, 1, growth=True), storage=bh.storage.Int()
+        bh.axis.Regular(10, 0, 1, growth=True), storage=bh.storage.Int64()
     )
 
     hist.fill(1.45)
@@ -111,7 +113,7 @@ def test_growing_histogram():
 
 def test_numpy_flow():
     h = bh.Histogram(
-        bh.axis.Regular(10, 0, 1), bh.axis.Regular(5, 0, 1), storage=bh.storage.Int()
+        bh.axis.Regular(10, 0, 1), bh.axis.Regular(5, 0, 1), storage=bh.storage.Int64()
     )
 
     for i in range(10):
@@ -135,7 +137,7 @@ def test_numpy_flow():
 
 def test_numpy_compare():
     h = bh.Histogram(
-        bh.axis.Regular(10, 0, 1), bh.axis.Regular(5, 0, 1), storage=bh.storage.Int()
+        bh.axis.Regular(10, 0, 1), bh.axis.Regular(5, 0, 1), storage=bh.storage.Int64()
     )
 
     xs = []
@@ -160,10 +162,10 @@ def test_numpy_compare():
 
 def test_project():
     h = bh.Histogram(
-        bh.axis.Regular(10, 0, 1), bh.axis.Regular(5, 0, 1), storage=bh.storage.Int()
+        bh.axis.Regular(10, 0, 1), bh.axis.Regular(5, 0, 1), storage=bh.storage.Int64()
     )
-    h0 = bh.Histogram(bh.axis.Regular(10, 0, 1), storage=bh.storage.Int())
-    h1 = bh.Histogram(bh.axis.Regular(5, 0, 1), storage=bh.storage.Int())
+    h0 = bh.Histogram(bh.axis.Regular(10, 0, 1), storage=bh.storage.Int64())
+    h1 = bh.Histogram(bh.axis.Regular(5, 0, 1), storage=bh.storage.Int64())
 
     for x, y in (
         (0.3, 0.3),
@@ -195,7 +197,7 @@ def test_sums():
 
 
 def test_int_cat_hist():
-    h = bh.Histogram(bh.axis.IntCategory([1, 2, 3]), storage=bh.storage.Int())
+    h = bh.Histogram(bh.axis.IntCategory([1, 2, 3]), storage=bh.storage.Int64())
 
     h.fill(1)
     h.fill(2)
